+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -751,7 +751,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K5ZT3CV9BJA1682124W3S9A4"),
+  _id: z.string().default("01K60CGQ6MNDX30BYC1PHKN43C"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -795,7 +795,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K5ZT3CV9BJA1682124W3S9A4"),
+  id: z.string().default("01K60CGQ6MNDX30BYC1PHKN43C"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,