+++ packages/orq-rc/src/models/operations/getagent.ts
@@ -946,7 +946,7 @@ export const GetAgentKnowledgeBases$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01K5ZT3CQCJNRJAK5JM21YBYM3"),
+  id: z.string().default("01K60CGQ2DXANBB2SN1A3NQ2CT"),
   knowledge_id: z.string(),
   configuration: z.union([
     z.lazy(() => GetAgentConfiguration2$inboundSchema),
@@ -973,7 +973,7 @@ export const GetAgentKnowledgeBases$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GetAgentKnowledgeBases
 > = z.object({
-  id: z.string().default("01K5ZT3CQCJNRJAK5JM21YBYM3"),
+  id: z.string().default("01K60CGQ2DXANBB2SN1A3NQ2CT"),
   knowledgeId: z.string(),
   configuration: z.union([
     z.lazy(() => GetAgentConfiguration2$outboundSchema),