+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -172,7 +172,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K5ZT3CV8RH1NSQQXZXZ9KF45"),
+  _id: z.string().default("01K60CGQ6M39ZNWYVDGTEW4VH2"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,
@@ -216,7 +216,7 @@ export const RetrieveDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K5ZT3CV8RH1NSQQXZXZ9KF45"),
+  id: z.string().default("01K60CGQ6M39ZNWYVDGTEW4VH2"),
   displayName: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$outboundSchema,