+++ packages/orq-rc/RELEASES.md
@@ -3849,3 +3849,13 @@ Based on:
 - [typescript v3.13.0-rc.36] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.36 - packages/orq-rc
+
+## 2025-09-25 12:12:05
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.628.4 (2.716.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.37] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.37 - packages/orq-rc